kgit: Avoid messing up a worktree when removing
"git worktree remove" does not check if it can actually remove the
entire worktree and will fail if there are files or directories owned
by someone else. In such a situation, the worktree metadata gets
removed, but the directory and the branch are still there, so they
would need to be disposed of manually.

Just let user know why kgit didn't commence with the worktree removal.